requirements.txt
gradio~=3.49.0
Pillow>=10.0.0
